+++ src/main/java/domain/participant/GameParticipant.java
@@ -1,5 +1,7 @@
-package domain;
+package domain.participant;
 
+import domain.card.Card;
+import domain.card.CardHand;
 import java.util.List;
 
 public class GameParticipant {
