+++ src/main/java/domain/participant/Player.java
@@ -1,4 +1,6 @@
-package domain;
+package domain.participant;
+
+import domain.card.CardHand;
 
 public class Player extends GameParticipant {
     public Player(String name, CardHand cardHand) {
