+++ src/main/java/domain/shuffler/Shuffler.java
@@ -1,5 +1,6 @@
-package domain;
+package domain.shuffler;
 
+import domain.card.Card;
 import java.util.List;
 
 public interface Shuffler {
