+++ src/test/java/domain/DealerTest.java
@@ -1,17 +1,22 @@
 package domain;
 
-import static domain.Rank.ACE;
-import static domain.Rank.FOUR;
-import static domain.Rank.NINE;
-import static domain.Rank.SIX;
-import static domain.Rank.TEN;
-import static domain.Rank.THREE;
-import static domain.Rank.TWO;
-import static domain.Suit.CLOVER;
-import static domain.Suit.DIAMOND;
-import static domain.Suit.HEART;
+import static domain.card.Rank.ACE;
+import static domain.card.Rank.FOUR;
+import static domain.card.Rank.NINE;
+import static domain.card.Rank.SIX;
+import static domain.card.Rank.TEN;
+import static domain.card.Rank.THREE;
+import static domain.card.Rank.TWO;
+import static domain.card.Suit.CLOVER;
+import static domain.card.Suit.DIAMOND;
+import static domain.card.Suit.HEART;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import domain.card.CardHand;
+import domain.card.Deck;
+import domain.game.GameResult;
+import domain.participant.Dealer;
+import domain.shuffler.RandomShuffler;
 import fixture.CardFixture;
 import java.util.Set;
 import org.junit.jupiter.api.DisplayName;
