+++ src/test/java/domain/GameTest.java
@@ -3,6 +3,12 @@ package domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import domain.card.Deck;
+import domain.game.Game;
+import domain.game.GameResult;
+import domain.participant.Dealer;
+import domain.participant.Player;
+import domain.shuffler.RandomShuffler;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
